(#507) Move unpack_args to interpreter unit
